+++ theme/source-more-technology-enterprise/page-about.php
@@ -1,4 +1,4 @@
 <?php
-/* Template Name: About Source More Technology */
+/* Template Name: About Source More */
 if (!defined('ABSPATH')) exit;
 require get_template_directory() . '/page-about-legacy.php';